Enhance settings menu at small sizes (#2350)
* Make the settings menu fullscreen below breakpoint 2
* Resume previous state when opening/closing the menu at small sizes
* Refactor settings code into it's own module
* Adjust max height at BP2 so the settings top bar doesn't get cut off at the low end of BP2
JW8-481
